\name{ExpMatrix}
\alias{ExpMatrix}
\docType{data}
\title{
The expression data
}
\description{
An example of gene expression data.
}
\usage{data("ExpMatrix")}
\format{
  The format is:
  
 num [1:200, 1:2150] 3.402 0.791 6.592 3.13 5.794 ...
 
 - attr(*, "dimnames")=List of 2
 
  ..$ : chr [1:200] "S1" "S2" "S3" "S4" ...
  
  ..$ : chr [1:2150] "8813" "2729" "1080" "5893" ...
}
\examples{
data(ExpMatrix)
}
\keyword{datasets}
