\name{ext2EntrezID}
\alias{ext2EntrezID}
\title{
Transform selection features into EntrezID
}
\description{
Transform selection features into EntrezID at the gene level.
}
\usage{
ext2EntrezID(x)
}
\arguments{
  \item{x}{
  Names of selected vars by \code{PCLasso} or \code{cv.PCLasso}.
}
}
\value{
Selected vars.
}
\author{
Wei Liu
}
