\name{ext2Group}
\alias{ext2Group}
\title{
Convert selected features into ComplexID
}
\description{
Convert selected features into ComplexID at protein complex level 
}
\usage{
ext2Group(x)
}
\arguments{
  \item{x}{
Names of selected vars by \code{PCLasso} or \code{cv.PCLasso}.
}
}
\value{
Selected protein complexes.
}
\author{
Wei Liu
}
