\name{PCGroup}
\alias{PCGroup}
\docType{data}
\title{
Protein complexes for "PCLasso"/"cv.PCLasso"
}
\description{
A list of protein complexes. The genes in each protein complex are represented 
by EntrezID, which are consistent with the gene names in \code{ExpMatrix}. 
}
\usage{data("PCGroup")}
\format{
  The format is:
  
List of 2417

 $ C_1   : chr [1:2] "604" "9759"
 
 $ C_2   : chr [1:2] "604" "10014"
 
 $ C_3   : chr [1:2] "604" "51564"
 
 $ C_4   : chr [1:4] "2033" "1387" "8850" "8202"
 
 $ C_11  : chr [1:2] "3257" "89781"
 
 $ C_12  : chr [1:3] "79803" "84343" "11234"
 
  [list output truncated]
}
\examples{
data(PCGroup)
}
\keyword{datasets}
